Rename SelectedSpeciesLozenge to SpeciesLozenge
