+++ src/shared/components/selected-species/SpeciesLozenge.scss
